small refactor of Telegram component
