+++ viseron/components/telegram/const.py
@@ -1,7 +1,9 @@
-"""Telegram-PTZ component constants."""
+"""Telegram component constants."""
+
+from typing import Final
 
 COMPONENT = "telegram"
-DESC_COMPONENT = "Telegram bot to control pan-tilt-zoom cameras."
+DESC_COMPONENT = "Telegram bot to control cameras."
 
 CONFIG_PTZ_COMPONENT = "ptz"
 
@@ -18,9 +20,17 @@ CONFIG_SEND_THUMBNAIL = "send_detection_thumbnail"
 CONFIG_SEND_VIDEO = "send_detection_video"
 CONFIG_SEND_MESSAGE = "send_detection_message"
 
+DEFAULT_TELEGRAM_USER_IDS: list[int] = []
+DEFAULT_SEND_THUMBNAIL: Final = False
+DEFAULT_SEND_VIDEO: Final = False
+DEFAULT_SEND_MESSAGE: Final = True
+DEFAULT_TELEGRAM_LOG_IDS: Final = False
+
 DESC_TELEGRAM_BOT_TOKEN = "Telegram bot token."
 DESC_TELEGRAM_CHAT_IDS = "List of chat IDs to send messages to."
-DESC_TELEGRAM_USER_IDS = "List of user IDs to accept commands from."
+DESC_TELEGRAM_USER_IDS = (
+    "List of user IDs to accept commands from. Only applicable in public chats."
+)
 DESC_TELEGRAM_LOG_IDS = "True if we should log the id of a user who was denied access."
 
 DESC_DETECTION_LABEL = "Label of the object to send notifications for."
